--- src/components/Table/EditableTable/types.tsx
@@ -1,8 +1,0 @@
-export interface Stock {
-  symbol: string;
-  buy_rate: number;
-  buy_date: string;
-  quantity: number;
-  close: number;
-  total: number;
-}
