+++ src/pages/CompareCompanyPage/index.tsx
@@ -2,9 +2,24 @@
  * Internal dependencies.
  */
 import { ComparisionTable } from "../../components";
+import useCompareCompanyPage from "./useCompareCompanyPage";
 
 const CompareCompanyPage = () => {
-  return <ComparisionTable />;
+  const {
+    companyOne,
+    companyTwo,
+    handleCompanyOneChange,
+    handleCompanyTwoChange,
+  } = useCompareCompanyPage();
+
+  return (
+    <ComparisionTable
+      companyOne={companyOne}
+      companyTwo={companyTwo}
+      handleCompanyOneChange={handleCompanyOneChange}
+      handleCompanyTwoChange={handleCompanyTwoChange}
+    />
+  );
 };
 
 export default CompareCompanyPage;
