+++ src/pages/PortfolioTrackerPage/useEditableTable.tsx
@@ -3,28 +3,115 @@
  */
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
-import { ReactElement, useState } from "react";
+import { ReactElement, useEffect, useState } from "react";
 import Alert from "@mui/material/Alert";
 
 /**
  * Internal dependencies.
  */
-import { type Stock } from "./types";
-import useEditableTableData from "./useEditableTableData";
 import {
   ADD_STOCK_PORTFOLIO_ENDPOINT,
   REMOVE_STOCK_PORTFOLIO_ENDPOINT,
   UPDATE_STOCK_PORTFOLIO_ENDPOINT,
-} from "../../../store/constant";
-import extractTextFromHTML from "../../../utilities/extractTextFromHTML";
+  GET_STOCK_PORTFOLIO_ENDPOINT,
+} from "../../store/constant";
+import useAppData from "../../store/useAppData";
+import extractTextFromHTML from "../../utilities/extractTextFromHTML";
 
-const useEditableTable = () => {
-  const { tableData } = useEditableTableData();
+export interface Stock {
+  symbol: string;
+  buy_rate: number;
+  buy_date: string;
+  quantity: number;
+  close: number;
+  total: number;
+}
+
+const useEditableTable = (
+  validationErrors: Record<string, string | undefined>,
+  setValidationErrors: (
+    validationErrors: Record<string, string | undefined>
+  ) => void
+) => {
   const navigate = useNavigate();
+  const { marketData } = useAppData();
+  const [tableData, setTableData] = useState([] as unknown as Stock[]);
 
   const [toastNotification, setToastNotification] = useState<ReactElement>(
     <></>
   );
+
+  const columns = [
+    {
+      accessorKey: "symbol",
+      header: "Symbol",
+      enableEditing: true,
+      size: 80,
+      muiEditTextFieldProps: {
+        required: true,
+        onFocus: () =>
+          setValidationErrors({
+            ...validationErrors,
+            symbol: undefined,
+          }),
+      },
+    },
+    {
+      accessorKey: "buy_date",
+      header: "Buy Date",
+      enableEditing: true,
+      size: 80,
+      muiEditTextFieldProps: {
+        required: false,
+        onFocus: () =>
+          setValidationErrors({
+            ...validationErrors,
+            buy_date: undefined,
+          }),
+      },
+    },
+    {
+      accessorKey: "buy_rate",
+      header: "Buy  Rate (in NRP)",
+      enableEditing: true,
+      muiEditTextFieldProps: {
+        required: true,
+        onFocus: () =>
+          setValidationErrors({
+            ...validationErrors,
+            buy_rate: undefined,
+          }),
+      },
+    },
+    {
+      accessorKey: "quantity",
+      header: "Kitta",
+      enableEditing: true,
+      muiEditTextFieldProps: {
+        required: true,
+        onFocus: () =>
+          setValidationErrors({
+            ...validationErrors,
+            kitta: undefined,
+          }),
+      },
+    },
+    {
+      accessorKey: "close",
+      header: "LTP",
+      enableEditing: false,
+    },
+    {
+      accessorKey: "total",
+      header: "Amount",
+      enableEditing: false,
+    },
+    {
+      accessorKey: "profit_loss",
+      header: "Profit / Loss",
+      enableEditing: false,
+    },
+  ];
 
   //CREATE hook (post new Stock to api)
   const useCreateStock = () => {
@@ -185,7 +272,52 @@ const useEditableTable = () => {
     };
   };
 
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await fetch(GET_STOCK_PORTFOLIO_ENDPOINT, {
+          method: "POST",
+          credentials: "include",
+        });
+
+        if (!response.ok) {
+          throw new Error("Failed to fetch data");
+        }
+
+        const json = await response.json();
+        const data = json.data;
+
+        const preparedData = data.map((item: Stock) => {
+          const close = Number(
+            marketData.find((newItem: Record<string, string>) => {
+              return newItem.symbol === item.symbol;
+            })?.close ?? 0
+          );
+          const total = item.quantity * close;
+          const profit_loss =
+            item.quantity * close - item.quantity * item.buy_rate;
+
+          return {
+            ...item,
+            close,
+            total,
+            profit_loss,
+          };
+        });
+
+        setTableData(preparedData);
+      } catch (error) {
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-expect-error
+        console.error("Error fetching data:", error.message);
+        setTableData([] as unknown as Stock[]);
+      }
+    };
+    fetchData();
+  }, [marketData]);
+
   return {
+    columns,
     toastNotification,
     useCreateStock,
     validateStock,
