+++ src/pages/WatchlistPage/index.tsx
@@ -1,11 +1,20 @@
+/**
+ * Internal Dependencies.
+ */
+import { EditableTable, MessageBox } from "../../components";
+import useAppData from "../../store/useAppData";
+
 const WatchlistPage = () => {
-  return (
-    <div className="bg-gray-200 dark:bg-gray-800 p-8 rounded-md shadow-md text-center">
-      <h1 className="text-2xl font-semibold mb-4">Under Construction</h1>
-      <p className="text-gray-600 dark:text-gray-400">
-        This component is currently under construction. Check back soon!
-      </p>
-    </div>
+  const { userData } = useAppData();
+
+  return userData.isLoggedIn ? (
+    <EditableTable />
+  ) : (
+    <MessageBox
+      message="Please login to use this service."
+      displayHomeBtn={true}
+      displayLoginBtn={true}
+    />
   );
 };
