+++ src/store/constant/index.ts
@@ -13,4 +13,7 @@ export {
   UPDATE_STOCK_PORTFOLIO_ENDPOINT,
   GET_STOCK_PORTFOLIO_ENDPOINT,
   REMOVE_STOCK_PORTFOLIO_ENDPOINT,
+  ADD_STOCK_WATCHLIST_ENDPOINT,
+  GET_STOCK_WATCHLIST_ENDPOINT,
+  REMOVE_STOCK_WATCHLIST_ENDPOINT,
 } from "./apiEndpoints";
