+++ src/store/constant/navItems.tsx
@@ -9,6 +9,41 @@ const servicesSubMenuItems = [
     description: "Overview of stock prices, indices, and market trends.",
   },
   {
+    label: "Portfolio Tracker",
+    href: "/portfolio-tracker",
+    icon: <div className="text-xl text-center">🔎</div>,
+    description:
+      "Effortlessly monitor and manage your investments with portfolio tracker.",
+  },
+  {
+    label: "Watchlist",
+    href: "/watchlist",
+    icon: <div className="text-xl text-center">📋</div>,
+    description:
+      "Watch over your favourite stocks by making a separate watchlist.",
+  },
+  {
+    label: "Compare Company",
+    href: "/compare-company",
+    icon: <div className="text-xl text-center">⚖</div>,
+    description:
+      "Easily compare key financial metrics between different companies.",
+  },
+  {
+    label: "Share Calculator",
+    href: "/share-calculator",
+    icon: <div className="text-4xl text-center">🖩</div>,
+    description:
+      "Get to know the actual figures when buying and selling shares.",
+  },
+  {
+    label: "Chart",
+    href: "/chart",
+    icon: <div className="text-xl text-center">📈</div>,
+    description:
+      "Visualize stock performance through dynamic and interactive charts.",
+  },
+  {
     label: "IPO Checker",
     href: "/ipo-checker",
     icon: <div className="text-xl text-center">✔️</div>,
@@ -28,41 +63,6 @@ const servicesSubMenuItems = [
     icon: <div className="text-xl text-center">📊</div>,
     description:
       "Receive actionable trading signals to assist in making strategic buy or sell decisions.",
-  },
-  {
-    label: "Watchlist",
-    href: "/watchlist",
-    icon: <div className="text-xl text-center">📋</div>,
-    description:
-      "Watch over your favourite stocks by making a separate watchlist.",
-  },
-  {
-    label: "Portfolio Tracker",
-    href: "/portfolio-tracker",
-    icon: <div className="text-xl text-center">🔎</div>,
-    description:
-      "Effortlessly monitor and manage your investments with portfolio tracker.",
-  },
-  {
-    label: "Chart",
-    href: "/chart",
-    icon: <div className="text-xl text-center">📈</div>,
-    description:
-      "Visualize stock performance through dynamic and interactive charts.",
-  },
-  {
-    label: "Compare Company",
-    href: "/compare-company",
-    icon: <div className="text-xl text-center">⚖</div>,
-    description:
-      "Easily compare key financial metrics between different companies.",
-  },
-  {
-    label: "Share Calculator",
-    href: "/share-calculator",
-    icon: <div className="text-4xl text-center">🖩</div>,
-    description:
-      "Get to know the actual figures when buying and selling shares.",
   },
 ];
